Made reading archives work.
The Darwin linker creates BSD ar archives, which (stupidly) always use
the "extended filename" method to store filenames.  For this reason I
implemented the BSD extended filename method in paxinc's ar_next.
Next to some fixes in the macho code not to free and unmap when this is
not desired, this yields in a working scanmacho on static archives.

The change has been checked with scanelf against a GNU ELF static
archive not to break anything.