Rewrite symbol matching code in scanelf.
The previous code was entirely broken when trying to match symbols in
unstripped binaries when giving multiple symbols as target.

The new code differs from the old one in quite a few ways:

- debug output when -g option is passed is disabled in the code
  (hacky, but still better than before!);

- regular expression matching is actually used when -g option is
  passed;

- by default, the symbol name is tested against the symbol name
  without version; no-version symbol name matching is possible by
  using the -g option and adding a final $;

- multiple symbols and single symbols are handled in the same way so
  that there is no more difference between them;

- the returned symbol name is the actual symbol name as found in the
  file, so includes the version when the file is not stripped.

While this means that there are some minimal differences between the
previous code, it's arguable that this code is less buggy and more
consistent than the one before.