fix possible memory read errors when walking arrays
the current code will always fetch the arr->eles[n] in array_for_each before doing the n < arr->num check.  gcc might optimize it such that the read occurs rather than delaying it until after the loop limit check, but it also might not.  at any rate, ASAN catches it and complains mightly.  this new method ends up wasting 1 pointer worth of memory, but we wont worry about 4 or 8 bytes per array as this code is not that critical.