move array_cnt check into array_for_each init
atm, if you try to use array_for_each or array_flatten_str on an array that has no members, you will get a segfault.  this is an easy rule to forget (and the current code does just that in at least one place), so move the array_cnt check into the init phase.  theres negligible code size impact so it should not be a big deal.