rewrite which() so that it works and does not break $PATH
we were walking the $PATH in reverse which it should have been forwards.  we were also modifying the pointer we got back from getenv() which meant any time we ran external code, $PATH would be truncated.  finally, we never actually checked the first element -- we would bail before we got a chance.