lddtree.py: dereference the interp symlink to handle desynced ABIs better
On aarch64/s390x, the interp is always placed in /lib/ while the libs are
normally placed in /lib64/.  If the interp is a symlink (which it normally
is here), then dereference it to locate the lib paths automatically.  This
also fixes up some bad lstrip logic in the /usr path expansion.

We could update lddtree.sh with a similar fix, but it would take more work
as it doesn't have the readlink helper to properly walk symlinks according
to an alternative root.  Plus, it already uses strings to try and extract
the paths the ldso was compiled against, and that seems to work in these
cases.

Reported-by: Jakub Čajka <jcajka@redhat.com>
URL: https://bugzilla.redhat.com/1162184