scanelf: add paren to quiet new gcc-5 warning
Newer gcc versions throw a warning on code that is written how we want
(comparing the result of !strcmp to a variable).  Add a set of paren to
make gcc happy.

scanelf.c: In function 'scanelf_file_needed_lib':
scanelf.c:1083:43: warning: logical not is only applied to the left hand side of comparison [-Wlogical-not-parentheses]
        if (!strcmp(find_lib_name, needed) == invert) \
                                           ^