lddtree: handle direct ldso linkage better
If the ELF we're showing is linked directly against the ELF which is used
as the ELF interp (i.e. the ldso), make sure we don't list it twice.

We want to see it "twice" in the pretty form:
$ ./lddtree.py /bin/bash
/bin/bash (interpreter => /lib/ld-linux-armhf.so.3)  <= here
    libreadline.so.6 => /lib/libreadline.so.6
    libncurses.so.5 => /lib/libncurses.so.5
    libc.so.6 => /lib/libc.so.6
    ld-linux-armhf.so.3 => /lib/ld-linux-armhf.so.3  <= here

But not in the list form (as that is used by scripts):
$ ./lddtree.py -l /bin/bash
/bin/bash
/lib/ld-linux-armhf.so.3  <= only here
/lib/libreadline.so.6
/lib/libncurses.so.5
/lib/libc.so.6

We also reconcile a few differences wrt interp display and the sh/py
implementations.

Reported-by: Tomasz Buchert <tomasz@debian.org>