debug: improve cleanup logic
Always set up the cleanup symbol and use it at C time rather than CPP.

We can delete the warning code since we clean those strings up now.

Add ASAN support so LSAN doesn't complain either.

Fix a bug in the leak checking when the -F flag is used -- normally we
don't allocate that string but set it to one of the argv constants.