lddtree.sh: fix interp handling when doing a full listing
On Debian multiarch systems, the ldso path is symlinked to a different
search path (and the main path isn't searched at all).  When listing the
deps of an ELF which links against the ldso, we end up showing the ldso
being loaded by two different paths:

$ ./lddtree.sh -a /bin/bash
bash => /bin/bash (interpreter => /lib64/ld-linux-x86-64.so.2)
    libtinfo.so.5 => /lib/x86_64-linux-gnu/libtinfo.so.5
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6
            ld-linux-x86-64.so.2 => /lib/x86_64-linux-gnu/ld-linux-x86-64.so.2
...

Fix the output to show that ld-linux-x86-64.so.2 was located at the same
path as the original interpreter as that better matches the runtime ldso
behavior.

$ ./lddtree.sh -a /bin/bash
bash => /bin/bash (interpreter => /lib64/ld-linux-x86-64.so.2)
    libtinfo.so.5 => /lib/x86_64-linux-gnu/libtinfo.so.5
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6
            ld-linux-x86-64.so.2 => /lib64/ld-linux-x86-64.so.2

Reported-by: Tomasz Buchert <tomasz@debian.org>