lddtree: handle exceptions thrown when parsing other ELFs
We have a top level exception handler for catching ELF errors, but if we
hit a problem with a dependency, we end up diagnosing it as a problem in
the first ELF.  e.g. If "foo" is linked against "libc.so", and "libc.so"
is a linker script, when we try to read the libc.so ELF, it fails, and
we end up saying:
  lddtree: warning: foo: Magic number does not match

This makes no sense because "foo" is a proper ELF with the right magic,
and we end up halting the whole parsing process.  Now we show:
  lddtree: warning: /usr/lib/libc.so: Magic number does not match
  foo (interpreter => /some/interp)
      libc.so => None