lddtree: respect (destination) root with --argv0 interp probing
This is important because the path in this func reflects where the
interpreter is going to be living, not where it's being copied from.
These are often the same paths (relatively speaking), but when we're
copying from a cross-compiler sysroot, or when using --libdir so the
target paths are rewritten, the interp path probably does not line
up with whatever happens to live in the active rootfs.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>