scanelf: allow user to choose whether to scan .symtab or .dynsym
This commit adds a new command line switch (-H/--how) that allows the
user to select whether scanelf should look for symbols in .symtab or
.dynsym.

This is done because the the sections contain different sets
of symbols (which somewhat intersect), and the names for the symbols
differ (.symtab includes the version suffix within the symbol itself),
which would cause scanelf to behave differently before and after
stripping the executable.

Additionally, it adds two modes of looking for symbols in .dynsym:
 - unversioned - look at the .dynsym entries as they are,
 - versioned - also look at .gnu.version to figure out the symbol
   version, and suffix the symbol name with it before matching,
   emulating the behavior of .symtab.

Note that this does not handle the case where the dynamic section needs
to be scanned to find the version information, as that case appears to
not be fully functional anyway (no DT_GNU_HASH support, and most ELFs
these days don't have DT_HASH). Consider it a TODO if need be :^).

Bug: https://bugs.gentoo.org/847493

Co-authored-by: Arsen Arsenović <arsen@aarsen.me>
Signed-off-by: Kacper Słomiński <kacper.slominski72@gmail.com>