github: switch the python checker CI to use something that is maintained
The python-lint action was last updated 3 years ago. It contains very
old versions of the lint tools, including pre-1.0 versions of mypy. It
doesn't allow installing dependencies for mypy to robustly typecheck.

Simply installing the tools ourselves and running them directly is
simpler, shorter, easier, and provides better linting analysis. So, do
so.

Another advantage is that we can now run all the linting tools, even if
an earlier one failed. As long as one step fails, the overall status is
still a failure.

In the process, we can remove a `# type: ignore` from lddtree.py;
argcomplete has typing info. So does pyelftools, but only in git master.